move constructor to sg.go
