+++ cmd/zoekt-sourcegraph-indexserver/main.go
@@ -9,7 +9,6 @@ import (
 	"flag"
 	"fmt"
 	"html/template"
-	"io"
 	"io/ioutil"
 	"log"
 	"math"
@@ -28,7 +27,6 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/hashicorp/go-retryablehttp"
 	"github.com/keegancsmith/tmpfriend"
 	"github.com/peterbourgon/ff/v3/ffcli"
 	"github.com/prometheus/client_golang/prometheus"
@@ -851,35 +849,6 @@ func startServer(conf rootConfig) error {
 
 	s.Run()
 	return nil
-}
-
-func newSourcegraphClient(rootURL *url.URL, hostname string, batchSize int) *sourcegraphClient {
-
-	client := retryablehttp.NewClient()
-	client.Logger = debug
-
-	// Sourcegraph might return an error message in the body if StatusCode==500. The
-	// default behavior of the go-retryablehttp client is to drain the body and not
-	// to propagate the error. Hence, we call ErrorPropagatedRetryPolicy instead of
-	// DefaultRetryPolicy and augment the error with the response body if possible.
-	client.CheckRetry = func(ctx context.Context, resp *http.Response, err error) (bool, error) {
-		shouldRetry, checkErr := retryablehttp.ErrorPropagatedRetryPolicy(ctx, resp, err)
-
-		if resp != nil && resp.StatusCode == http.StatusInternalServerError {
-			if b, e := io.ReadAll(resp.Body); e == nil {
-				checkErr = fmt.Errorf("%w: body=%q", checkErr, string(b))
-			}
-		}
-
-		return shouldRetry, checkErr
-	}
-
-	return &sourcegraphClient{
-		Root:      rootURL,
-		Client:    client,
-		Hostname:  hostname,
-		BatchSize: batchSize,
-	}
 }
 
 func newServer(conf rootConfig) (*Server, error) {
